+++ classes/MyPoint.java
@@ -7,8 +7,8 @@
 package classes;
 
 public class MyPoint {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     // Default Constructor
     public MyPoint() {
@@ -16,28 +16,28 @@ public class MyPoint {
         this.y = 0;
     }
 
-    public MyPoint(int x, int y) {
+    public MyPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void setXY(int x, int y) {
+    public void setXY(double x, double y) {
         this.x = x;
         this.y = y;
     }
